lightnvm: do not reserve gc blocks on gennvm
GC is strictly handled by the target. The media manager should therefore
not reserve blocks for GC or any other purpose.

In order to support targets implementing atomic get_block based on the
number of free blocks in the lun, introduce an unlocked version og
get_block.

After this change, flags are not needed for get_blk. Adapt functions
and rrpc.

Signed-off-by: Javier González <javier@cnexlabs.com>