lightnvm-dflash: Removed dflash_lun and added attributes to dflash.
The `dflash_lun` structure had variables such as blocks, nr_blocks, and
nr_free_blocks, however, they were not maintained. E.g. nr_free_blocks was
incremented and decremented but initial values did not align with those in the
media manager. Similarly, the dflash_lun.blocks were allocated but never used.

With this change, dflash->luns is an array of pointers to the media-manager
nvm_luns. Responsibility of accounting the nr_blocks and nr_free_blocks, etc is
thus delegated to the media manager.

The wrapping / indirection of dflash_lun might be required but this change
removes it, since the current state it does not do anything.